+++ src/main/java/de/uib/configed/guidata/InstallationStateTableModel.java
@@ -1254,7 +1254,7 @@ public class InstallationStateTableModel extends AbstractTableModel implements C
 	public void setValueAt(Object value, int row, int col) {
 		Logging.debug(this, " actualProduct " + actualProduct + ", set value at " + row + ", " + col);
 		changeValueAt(value, originRow(row), col);
-		fireTableCellUpdated(originRow(row), col);
+		fireTableCellUpdated(row, col);
 	}
 
 	protected void changeValueAt(Object value, int row, int col) {
@@ -1286,7 +1286,6 @@ public class InstallationStateTableModel extends AbstractTableModel implements C
 		if (!retrieveValue.equals(value)) {
 			changeValue(value, col);
 		}
-
 	}
 
 	private void changeValue(Object value, int col) {