+++ src/main/java/de/uib/opsidatamodel/serverdata/dataservice/GroupDataService.java
@@ -25,7 +25,6 @@ import de.uib.opsidatamodel.serverdata.CacheIdentifier;
 import de.uib.opsidatamodel.serverdata.CacheManager;
 import de.uib.opsidatamodel.serverdata.OpsiServiceNOMPersistenceController;
 import de.uib.opsidatamodel.serverdata.RPCMethodName;
-import de.uib.utilities.datastructure.StringValuedRelationElement;
 import de.uib.utilities.logging.Logging;
 import utils.Utils;
 
@@ -377,7 +376,7 @@ public class GroupDataService {
 		return result;
 	}
 
-	public boolean addGroup(StringValuedRelationElement newgroup, boolean isHostGroup) {
+	public boolean addGroup(Map<String, String> newgroup, boolean isHostGroup) {
 		if (!userRolesConfigDataService.hasServerFullPermissionPD()) {
 			return false;
 		}